requirements.txt
apipkg==1.5
atomicwrites==1.3.0
attrs==19.1.0
allure_python_commons==2.9.45
allure-pytest==2.9.45
certifi==2019.6.16
chardet==3.0.4
execnet==1.7.1
fixtures==3.0.0
glob2==0.7
idna==2.8
Mako==1.1.0
MarkupSafe==1.1.1
more-itertools==7.0.0
packaging==19.2
parse==1.11.1
parse-type==0.4.2
pi==0.1.2
pipenv==2022.1.8
pluggy==0.13.0
py==1.10.0
pyparsing==2.4.2
pytest==7.0.1
pytest-bdd==5.0.0
pytest-forked==1.1.3
pytest-html==2.0.0
pytest-metadata==1.8.0
pytest-xdist==1.31.0
selenium==4.1.3
webdriver_manager==3.5.2